adventures/30_vertex_pipelines/app/requirements.txt
google-cloud-aiplatform==1.49.0
kfp==2.6.0
google-cloud-storage==2.16.0
